+++ src/backend/joanie/tests/core/test_models_contract_definition.py
@@ -37,3 +37,31 @@ class ContractDefinitionModelTestCase(TestCase):
         result = contract_definition.get_body_in_html()
 
         self.assertEqual(result, expected_html)
+
+    def test_models_contract_definition_get_appendix_in_html_with_empty_appendix(self):
+        """
+        We should get an empty string in return when we call the get appendix in HTML method
+        if the contract definition appendix is empty.
+        """
+        contract_definition = factories.ContractDefinitionFactory(
+            title="Contract Definition 1", appendix=""
+        )
+
+        result = contract_definition.get_appendix_in_html()
+
+        self.assertEqual(result, "")
+
+    def test_models_contract_definition_get_appendix_in_html(self):
+        """
+        We should get the HTML format of the markdown content of the contract definition appendix
+        when we call get appendix in HTML method.
+        """
+        content = "## The student must have a computer to follow the course"
+        expected_html = "<h2>The student must have a computer to follow the course</h2>"
+        contract_definition = factories.ContractDefinitionFactory(
+            title="Contract Definition 1", appendix=content
+        )
+
+        result = contract_definition.get_appendix_in_html()
+
+        self.assertEqual(result, expected_html)