+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- Add `appendix` field on `ContractDefinition` model
+
 ## [2.8.0] - 2024-10-16
 
 ### Added